requirements-dev.txt
-r requirements.txt

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-flask==1.3.0

# Code Quality
black==23.12.1
flake8==6.1.0
mypy==1.7.1

# Type stubs
types-Flask==1.1.6
